requirements.txt
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
flagpy==1.0.3
folium==0.12.1
geographiclib==1.52
geopy==2.2.0
html2image==2.0.1
idna==3.3
ImageHash==4.2.1
imageio==2.10.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.3
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tifffile==2021.10.12
tk==0.1.0
urllib3==1.26.7
